fixed bug with deleting movie

+++ src/main/java/com/example/kinomaniak/controller/ManagerMovieViewController.java
@@ -118,8 +118,11 @@ public class ManagerMovieViewController {
     @FXML
     public void submitDeletingMovie() {
         Movie movieToDelete = movies.stream().filter(movie -> (movie.getTitle().equals(currentMovieTitleToRemove.getValue()))).findAny().get();
-        managerService.removeMovie(movieToDelete);
-        movies.remove(movieToDelete);
+        if(!managerService.movieIsBeingScreened(movieToDelete)){
+            managerService.removeMovie(movieToDelete);
+            movies.remove(movieToDelete);
+        }
+
     }
 
     @FXML
